flytting av route + endring skrivemåte
