+++ web/src/main/kotlin/no/kartverket/matrikkel/bygning/routes/v1/intern/bygning/BygningRoutes.kt
@@ -81,75 +81,3 @@ fun Route.bygningRouting(
         }
     }
 }
-
-fun Route.bruksenhetRouting(bygningService: BygningService) {
-    route(
-        "{bruksenhetId}",
-        {
-            request {
-                pathParameter<String>("bruksenhetId") {
-                    required = true
-                }
-            }
-            response {
-                code(HttpStatusCode.NotFound) {
-                    description = "Fant ikke bruksenhet med gitt bruksenhetId"
-                }
-            }
-        },
-    ) {
-        get(
-            {
-                summary = "Hent en bruksenhet"
-                description = "Hent en bruksenhet"
-                response {
-                    code(HttpStatusCode.OK) {
-                        body<BruksenhetInternResponse> {
-                            description = "Bruksenhet"
-                        }
-                        description = "Bruksenheten finnes og ble hentet"
-                    }
-                }
-            },
-        ) {
-            val bruksenhetId = call.parameters.getOrFail("bruksenhetId").toLong()
-
-            val (status, body) = bygningService.getBruksenhetByBubbleId(
-                bruksenhetBubbleId = bruksenhetId,
-            ).mapBoth(
-                success = { HttpStatusCode.OK to it.toBruksenhetResponse() },
-                failure = ::domainErrorToResponse,
-            )
-
-            call.respond(status, body)
-        }
-
-        route("egenregistrert") {
-            get(
-                {
-                    summary = "Hent egenregistrert data for en bruksenhet"
-                    description = "Hent egenregistrert data en bruksenhet"
-                    response {
-                        code(HttpStatusCode.OK) {
-                            body<BruksenhetSimpleResponse> {
-                                description = "Bruksenhet med én datakilde - kun egenregistrerte data"
-                            }
-                            description = "Bruksenheten finnes og ble hentet"
-                        }
-                    }
-                },
-            ) {
-                val bruksenhetId = call.parameters.getOrFail("bruksenhetId").toLong()
-
-                val (status, body) = bygningService.getBruksenhetByBubbleId(
-                    bruksenhetBubbleId = bruksenhetId,
-                ).mapBoth(
-                    success = { HttpStatusCode.OK to it.toBruksenhetSimpleResponseFromEgenregistrertData() },
-                    failure = ::domainErrorToResponse,
-                )
-
-                call.respond(status, body)
-            }
-        }
-    }
-}
